Description
Branching with if , else and elif. One of the most powerful features of programming languages is branching: the ability to make decisions and execute a different set of statements based on whether one or more conditions are true.
Apart form that loops are important in Python or in any other programming language as they help us  to execute a block of code repeatedly.
In this repository we will learn about these things.