Description
A while loop is a control flow statement which repeatedly executes a block of code until the condition is satisfied. It stops executing the block only if the condition fails. One should use a 'while' loop when one needs to perform a repeated operation, but doesn't know in advance how many iterations would be needed.