Description
The for loop in Python is used to iterate over a sequence (like a list, tuple, or string) or other iterable objects. Iterating over a sequence means going through each element one by one. In this article, we're going to describe how to iterate over a python list, using the built-in function for.